chore: Start to mock conversations
